Fixed test cases to line up with the justification.
Note: There is still a pathologic case, which the programmer can change a local variable within a method that called in the same line that use this local variable.
"setNum() * num;" kind of usage.

If we decide that this is a normal programming behavior, than the concept of shortest operand will lost.