Fixed the bug in the "whole project cleanup" option - because we are going to use a lot of progress bars here is why I believe the bug occurred - the previous process was:
1. Open a progress bar window
2. Call a new thread for execution
3. Call compilationUnits()
4. Crash! Why?

(4) Crash because we are using:
getCompilationUnit(getCurrentWorkbenchWindow().getActivePage().getActiveEditor())

But! I believe that the current window is no longer the editor - but the progress bar, therefore there is no ActivePage/ActiveEditor and we fail.

< The bottom line >
If we don't have a better methodology to get the compilation unit we have to remember not to use getCompilationUnit (or related functions) within an active progress bar or any other active window.