Fixed issue #57
There's still a tiny bug - the short name is added an "s" if the type contains multiple values. A type is considered to contain multiple values if it is variadic, an array or a parameterized type, which is mostly correct, however there are parameterized types that are not collections. Nonetheless, since it's problematic to determine whether a type is a collection, and since most parameterized types in use are collections, it was left as is.